third_party/WebKit/LayoutTests/fast/dom/shadow/shadowroot-type.html

<!DOCTYPE html>
<html>
<body>
<script src="../../../resources/js-test.js"></script>

<div id="container">
    <div id="host"></div>
    <input id="input" type="text">
</div>
<pre id="console"></pre>

<script>
description("This test ensures ShadowRootType is correctly reflected");

var shadowRootForDiv = host.createShadowRoot();
var shadowRootForInput = internals.shadowRoot(input);

shouldBe('"V0ShadowRoot"', 'internals.shadowRootType(shadowRootForDiv)');
shouldBe('"UserAgentShadowRoot"', 'internals.shadowRootType(shadowRootForInput)');

container.innerHTML = "";

finishJSTest();
</script>
</body>
</html>
